requirements.txt
#
# This file is autogenerated by pip-compile
# To update, run:
#
#    pip-compile
#
certifi==2020.12.5
    # via requests
chardet==4.0.0
    # via requests
flake8==3.9.1
    # via -r requirements.in
idna==2.10
    # via requests
importlib-metadata==3.10.1
    # via flake8
marshmallow-dataclass==8.4.1
    # via -r requirements.in
marshmallow==3.11.1
    # via
    #   -r requirements.in
    #   marshmallow-dataclass
mccabe==0.6.1
    # via flake8
mypy-extensions==0.4.3
    # via
    #   mypy
    #   typing-inspect
mypy==0.812
    # via -r requirements.in
pycodestyle==2.7.0
    # via flake8
pyflakes==2.3.1
    # via flake8
pyyaml==5.4.1
    # via -r requirements.in
requests==2.25.1
    # via -r requirements.in
typed-ast==1.4.3
    # via mypy
typing-extensions==3.7.4.3
    # via
    #   importlib-metadata
    #   mypy
    #   typing-inspect
typing-inspect==0.6.0
    # via marshmallow-dataclass
urllib3==1.26.4
    # via requests
zipp==3.4.1
    # via importlib-metadata
